+++ react-components/src/architecture/base/concreteCommands/pointCloud/PointCloudFilterCommand.ts
