+++ react-components/src/architecture/base/concreteCommands/pointCloud/SetPointColorTypeCommand.ts
