+++ react-components/src/architecture/base/concreteCommands/pointCloud/SetPointShapeCommand.ts
