+++ react-components/src/architecture/base/concreteCommands/pointCloud/SetPointSizeCommand.ts
